demo/requirements.txt
chromadb>=0.5.0
together>=1.0.0
gradio>=4.0.0
python-dotenv>=1.0.0
